requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# IA
google-generativeai==0.3.2
pandas==2.1.3
numpy==1.25.2

# Auth y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilidades
python-dotenv==1.0.0

email-validator==2.1.0